Tidy up the urls/views for paas_wl.bk_app.processes.
