--- apiserver/paasng/paas_wl/bk_app/processes/urls_enduser.py
@@ -1,40 +1,0 @@
-# -*- coding: utf-8 -*-
-"""
-TencentBlueKing is pleased to support the open source community by making
-蓝鲸智云 - PaaS 平台 (BlueKing - PaaS System) available.
-Copyright (C) 2017 THL A29 Limited, a Tencent company. All rights reserved.
-Licensed under the MIT License (the "License"); you may not use this file except
-in compliance with the License. You may obtain a copy of the License at
-
-    http://opensource.org/licenses/MIT
-
-Unless required by applicable law or agreed to in writing, software distributed under
-the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND,
-either express or implied. See the License for the specific language governing permissions and
-limitations under the License.
-
-We undertake not to change the open source license (MIT license) applicable
-to the current version of the project delivered to anyone in the future.
-"""
-from paasng.utils.basic import make_app_pattern_with_applications_prefix, re_path
-
-from . import views_enduser
-
-urlpatterns = [
-    # TODO: replace `api.processes` with `api.processes.update`
-    re_path(
-        make_app_pattern_with_applications_prefix(r"/processes/$"),
-        views_enduser.ProcessesViewSet.as_view({"post": "update"}),
-        name="api.processes",
-    ),
-    re_path(
-        make_app_pattern_with_applications_prefix(r"/processes/list/$"),
-        views_enduser.ListAndWatchProcsViewSet.as_view({"get": "list"}),
-        name="api.list_processes",
-    ),
-    re_path(
-        make_app_pattern_with_applications_prefix(r"/processes/watch/$"),
-        views_enduser.ListAndWatchProcsViewSet.as_view({"get": "watch"}),
-        name="api.watch_processes",
-    ),
-]
